requirements.txt
streamlit==1.26.0
google-generativeai==0.1.6
requests==2.31.0
folium==0.14.0
streamlit-folium==0.12.0
